requirements.txt
pymupdf==1.23.7
python-docx==0.8.11
pdfplumber==0.10.2
pyyaml==6.0.1
qdrant-client==1.6.4
sentence-transformers==2.2.2
huggingface-hub==0.20.3
openai==1.12.0
llama-cpp-python==0.2.23
pillow==10.1.0
numpy==1.26.3
python-dotenv==1.0.0
